requirements.txt
psycopg2==2.6.1
Django==1.7.11
Jinja2==2.8
jingo==0.8.2
Pygments==2.1.3
python-memcached==1.58
celery==3.1.23
django-celery==3.1.17
django-celery-email==1.1.4
Markdown==2.6.6
sorl-thumbnail==12.3
Pillow==3.2.0
mincss==0.11.1
django-fancy-cache==0.8.0
cssmin==0.2.0
cssselect==0.9.1
BeautifulSoup==3.2.1
hiredis==0.2.0
redis==2.10.5
zope.structuredtext==4.1.0
Unidecode==0.04.19
requests==2.10.0
six==1.10.0
django-nose==1.4.3
nose==1.3.7
amqp==1.4.9
anyjson==0.3.3
lxml==3.6.0
subprocess32==3.2.7
django-semanticui-form==0.0.1
django-pipeline==1.5.4
feedparser==5.2.1
pyquery==1.2.13
xmltodict==0.10.1
pyOpenSSL==16.0.0
ndg-httpsclient==0.4.0
pyasn1==0.1.9
jsonfield==1.0.3
awesome-slugify==1.6.5
pygments-lexer-babylon==0.12.0
optisorl==0.2.1
pytz==2016.4
